Attempt to add chart visual for voting trend
Most of my time was spent trying to add data with JS objects to this map
my goal was to try to iterate through the items object and match the
state with the states object with the data in the data.js file.
Then populate the winning candidate data into the tool tip.

I ran into a few hurdles here not knowing how to plot the objects and
how to manipulate the data. I think with a bit more time I would have
figured it out.