Attempt to serialize form submission
- created form, attempted to serialize but on submission the
.serialize() method throws an error.
- I am not familiar with serialization and I attempted to do some
research on how to do that. I didn't find much and the little that I
found was on the jquery method.

In hindsight I, should have maybe started here or planned out some
research time for new concepts and/or  asked for further clarification

Moved form data into main container